+++ processor/spark-run.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
-product=$1
+product_id=$1
 event_date=$2
 action=$3
 models=$4
 
 # SPARK_LOCAL_HOSTNAME=localhost spark-submit \
-#   --name "minerva::$product::$event_date::$action::$models" \
+#   --name "minerva::$product_id::$event_date::$action::$models" \
 #   --master local[*] \
-python3 src/app.py product=$product event_date=$event_date action=$action models=$models config_dir=../products
+python3 src/app.py product_id=$product_id event_date=$event_date action=$action models=$models config_dir=../products