+++ processor/src/app.py
@@ -8,7 +8,7 @@ if len(sys.argv) < 2:
   raise ValueError('No arguments passed.')
 
 settings = JobSettings(**string_utils.convert_args_to_dict(sys.argv[1:]))
-product_settings = ProductSettings(**file_utils.load_json(f'{settings.config_dir}/{settings.product}/profile.json'))
+product_settings = ProductSettings(**file_utils.load_json(f'{settings.config_dir}/{settings.product_id}/profile.json'))
 
 if settings.connector == 'spark': spark_runner.run(product_settings, settings)
 else: print(f'Unknown connector: {settings.connector}')